Added PHPDocs to all my Controllers
